refactor: 요구사항 테스트 각 목적에 맞게 클래스 분리
