+++ src/test/java/roomescape/reservation/repository/DataSourceTest.java
@@ -1,9 +1,7 @@
-package roomescape;
-
+package roomescape.reservation.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.lang.reflect.Field;
 import java.sql.Connection;
 import java.sql.SQLException;
 import org.junit.jupiter.api.DisplayName;
@@ -14,17 +12,14 @@ import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.global.config.TestConfig;
-import roomescape.reservation.controller.ReservationController;
 
 @Import(TestConfig.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class GlobalIntegrationTest {
+class DataSourceTest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
-    @Autowired
-    private ReservationController reservationController;
 
     @DisplayName("연결된 DB 커넥션을 확인한다")
     @Test
@@ -36,21 +31,6 @@ class GlobalIntegrationTest {
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-    }
-
-    @DisplayName("Controller에서 JdbcTemplate의 의존을 분리한다")
-    @Test
-    void jdbcTemplate_inject_test() {
-        boolean isJdbcTemplateInjected = false;
-
-        for (Field field : reservationController.getClass().getDeclaredFields()) {
-            if (field.getType().equals(JdbcTemplate.class)) {
-                isJdbcTemplateInjected = true;
-                break;
-            }
-        }
-
-        assertThat(isJdbcTemplateInjected).isFalse();
     }
 
 }
